Allow large IDs in the database (#75)

+++ athena/athena/models/db_exercise.py
@@ -1,11 +1,11 @@
-from sqlalchemy import Column, Integer, String, Float, JSON, Enum as SqlEnum
+from sqlalchemy import Column, BigInteger, String, Float, JSON, Enum as SqlEnum
 
 from athena.schemas import ExerciseType
 from .model import Model
 
 
 class DBExercise(Model):
-    id = Column(Integer, primary_key=True, index=True, nullable=False)
+    id = Column(BigInteger, primary_key=True, index=True, nullable=False)
     title = Column(String, index=True, nullable=False)
     type = Column(SqlEnum(ExerciseType), index=True, nullable=False)
     max_points = Column(Float, index=True, nullable=False)
